requirements.txt
scikit-surprise
pandas
numpy
fastapi
pydantic
sqlalchemy
jinja2
aiofiles